% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_to_tissue}
\alias{aggregate_to_tissue}
\title{Aggregate expression to tissue level}
\usage{
aggregate_to_tissue(se, tissue_column, aggregation_method = "median")
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object with expression data in
TPM (or normalized by library size) and a \code{colData} slot.}

\item{tissue_column}{Character indicating the name of the column
in \code{colData(se)} containing tissue information for each sample.}

\item{aggregation_method}{Character indicating the method to use to
aggregate samples from the same tissue before calculating \eqn{\tau}
(see details below). One of "median" or "mean". Default: "median".}
}
\value{
A gene expression matrix with genes in rows and tissues in columns.
}
\description{
Aggregate expression to tissue level
}
\examples{
# Load data
data(ath_se)
se <- ath_se

# Simulate different tissues
SummarizedExperiment::colData(se)$tissue <- c(rep("seed", 3), rep("leaf", 3))

# Aggregate expression
aexp <- aggregate_to_tissue(se, "tissue")
}
