% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ath_dups}
\alias{ath_dups}
\title{Duplicate gene pairs in the genome of Arabidopsis thaliana}
\format{
A data frame with the following variables:
\itemize{
\item{\emph{dup1} Character, duplicated gene 1.}
\item{\emph{dup2} Character, duplicated gene 2.}
\item{\emph{type} Factor, duplication mode.}
}
}
\usage{
data(ath_dups)
}
\description{
Duplicate gene pairs were identified and classified by duplication
mode with the \strong{doubletrouble} package.
}
\examples{
data(ath_dups)
}
\keyword{datasets}
