% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ath_se}
\alias{ath_se}
\title{Expression data for experiment SRP213876}
\format{
A \code{SummarizedExperiment} object with an assay named \strong{TPM}
and three colData columns:
\itemize{
\item{\emph{experiment} Character, experiment accession.}
\item{\emph{tissue} Character, tissue name.}
\item{\emph{treatment} Character, treatment.}
}
}
\usage{
data(ath_se)
}
\description{
Expression data (in TPM) and sample metadata were obtained from
Refine Bio.
}
\examples{
data(ath_se)
}
\keyword{datasets}
