% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_cor}
\alias{calculate_cor}
\title{Calculate correlation coefficients and permutation P-values for gene pairs}
\usage{
calculate_cor(se, dups, assay_name = NULL, method = "spearman")
}
\arguments{
\item{se}{A \code{SummarizedExperiment} or \code{SingleCellExperiment} or
\code{SpatialExperiment} object with expression data.}

\item{dups}{A data frame of duplicate pairs, with duplicated genes
for each pair in columns 1 and 2. Additional columns are allowed, and will
be ignored if present.}

\item{assay_name}{Character specifying the name of the assay to use. If
NULL, the first assay will be extracted. Default: NULL.}

\item{method}{Character indicating which correlation coefficient to compute.
One of 'pearson', 'spearman', or 'kendall'. Default: 'spearman'.}
}
\value{
A data frame exactly as the one passed in \strong{dups},
but with an extra column named \code{cor} (numeric) with correlation
coefficients. Pairs will have NA in this column if i. at least of the
genes in the pair is not present in the assay slot of \strong{se}; or ii.
at least of the genes in the pair has a standard deviation of 0.
}
\description{
Calculate correlation coefficients and permutation P-values for gene pairs
}
\examples{
# Load data
data(ath_se)
data(ath_dups)

# Subset data
dups <- ath_dups[ath_dups$type == "TD", ]
se <- ath_se

# Calculate correlation
cor_df <- calculate_cor(se, dups, method = "pearson")
}
