% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_tau}
\alias{calculate_tau}
\title{Calculate \eqn{\tau} (Tau) index of tissue specificity}
\usage{
calculate_tau(aggregated_exp, log = TRUE)
}
\arguments{
\item{aggregated_exp}{A gene expression matrix with genes in rows
and tissues in columns as returned by \code{aggregate_to_tissue()}.}

\item{log}{Logical indicating whether the matrix of aggregated
gene expression per tissue should be log-transformed
using \eqn{log_2(x+1)} before calculating \eqn{\tau}.}
}
\value{
A data frame with the following variables:
\itemize{
\item{\emph{gene} Character, gene ID.}
\item{\emph{tau} Numeric, \eqn{\tau} index of tissue specificity.}
\item{\emph{class} Factor, expression-based class. One of 'Null',
'Weak', 'Broad', 'Specific' (see Details field).}
}.
}
\description{
Calculate \eqn{\tau} (Tau) index of tissue specificity
}
\details{
Besides calculating \eqn{\tau} indices, this function classifies genes into
expression-based classes, namely:
\itemize{
\item{\strong{Null} Expression <1 in all tissues.}
\item{\strong{Weak} Expression <5 in all tissues.}
\item{\strong{Broad} \eqn{\tau} <0.85.}
\item{\strong{Specific} \eqn{\tau} >=0.85.}
}
}
\examples{
# Load data
data(ath_se)
se <- ath_se

# Simulate different tissues
SummarizedExperiment::colData(se)$tissue <- c(rep("seed", 3), rep("leaf", 3))

# Aggregate expression to tissue level
aexp <- aggregate_to_tissue(se, "tissue")

# Calculate Tau
tau_df <- calculate_tau(aexp)
}
