% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compare_coex_me}
\alias{compare_coex_me}
\title{Compare coexpression module eigengenes for diverged duplicate pairs}
\usage{
compare_coex_me(module_comp, ME, cor_method = "spearman")
}
\arguments{
\item{module_comp}{A data frame with module assignment comparison
between duplicate pairs as returned by \code{compare_coex_modules()}.}

\item{ME}{A matrix or data frame of module eigengenes, with samples in rows
and eigengenes in columns. This can be obtained, for instance,
with \code{BioNERO::exp2gcn()}.}

\item{cor_method}{Character indicating the correlation method to use
to calculate pairwise correlations between module eigengenes.
Default: "spearman".}
}
\value{
A data frame with duplicate pairs (and any additional columns
with metadata on pairs) and a column named \emph{ME_cor} with correlation
between eigengenes of the modules where each gene is.
}
\description{
Compare coexpression module eigengenes for diverged duplicate pairs
}
\details{
This function aims to derive a continuous measure from a binary variable
(preserved vs diverged modules for genes in a duplicate pair). For that,
this function calculates correlation coefficients between module eigengenes
for diverged gene pairs. Such correlation coefficients can be interpreted
as a measure of how much genes in a pair are diverged (i.e., if they
are in very different modules, or in quite similar modules).
}
\examples{
data(ath_dups)
data(genes_and_modules)
data(ME)
dups <- ath_dups
genes_modules <- genes_and_modules

# Compare
module_comp <- compare_coex_modules(dups, genes_and_modules)
comp <- compare_coex_me(module_comp, ME)
}
