% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compare_coex_modules}
\alias{compare_coex_modules}
\title{Compare coexpression module assignment for gene pairs}
\usage{
compare_coex_modules(dups, genes_modules)
}
\arguments{
\item{dups}{A data frame of duplicate pairs, with duplicated genes
for each pair in columns 1 and 2. Additional columns are allowed, and will
be ignored if present.}

\item{genes_modules}{A 2-column data frame with genes in column 1
and coexpression modules to which they belong in column 1.
This can be obtained, for instance, with \code{BioNERO::exp2gcn()}.}
}
\value{
A data frame with the same columns as the input to \strong{dups},
but with the following extra columns:
\itemize{
\item{\emph{module_dup1} Character, name of module where gene 1 is.}
\item{\emph{module_dup2} Character, name of module where gene 2 is.}
\item{\emph{module_preservation} Factor, classification of the
gene pair based on module preservation. Levels include
'preserved' (both genes in the same module), 'diverged' (genes
in different modules), 'one_absent' (one of the genes is not any
module), and 'both_absent' (both genes are not present in any module).}
}
}
\description{
Compare coexpression module assignment for gene pairs
}
\examples{
data(ath_dups)
data(genes_and_modules)
dups <- ath_dups
genes_modules <- genes_and_modules

# Compare
cm <- compare_coex_modules(dups, genes_and_modules)
}
