% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compare_tau}
\alias{compare_tau}
\title{Compare \eqn{\tau} indices for gene pairs}
\usage{
compare_tau(tau_df, dups)
}
\arguments{
\item{tau_df}{A 3-column data frame with \eqn{\tau} indices and expression
classes for each gene as returned by \code{calculate_tau()}.}

\item{dups}{A data frame of duplicate pairs, with duplicated genes
for each pair in columns 1 and 2. Additional columns are allowed, and will
be ignored if present.}
}
\value{
A data frame with the same columns as the input to \strong{dups},
but with the following extra columns:
\itemize{
\item{\emph{tau_dup1} Numeric, \eqn{\tau} index of gene 1.}
\item{\emph{class_dup1} Factor, expression-based class of gene 1.}
\item{\emph{tau_dup2} Numeric, \eqn{\tau} index of gene 2.}
\item{\emph{class_dup2} Factor, expression-based class of gene 2.}
\item{\emph{tau_diff} Numeric, absolute \eqn{\tau} difference between
duplicate pairs.}
\item{\emph{class_preserved} Logical indicating if expression-based class
is the same for both genes in a duplicate pair.}
}
}
\description{
Compare \eqn{\tau} indices for gene pairs
}
\examples{
# Load data
data(ath_se)
data(ath_dups)
dups <- ath_dups
se <- ath_se

# Simulate different tissues
SummarizedExperiment::colData(se)$tissue <- c(rep("seed", 3), rep("leaf", 3))

# Aggregate expression to tissue level
aexp <- aggregate_to_tissue(se, "tissue")

# Calculate Tau
tau_df <- calculate_tau(aexp)

# Compare Tau
ct <- compare_tau(tau_df, dups)
}
