% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdiva-package.R
\docType{package}
\name{exdiva-package}
\alias{exdiva}
\alias{exdiva-package}
\title{exdiva: A metric set to compute expression divergence between gene pairs}
\description{
exdiva provides users with a set of metrics to calculate expression divergence between gene pairs. The metrics include correlation coefficients, the Tau index of tissue specificity, and comparison of coexpression module assignment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almeidasilvaf/exdiva}
  \item Report bugs at \url{https://support.bioconductor.org/t/exdiva}
}

}
\author{
\strong{Maintainer}: Fabrício Almeida-Silva \email{fabricio_almeidasilva@hotmail.com} (\href{https://orcid.org/0000-0002-5314-2964}{ORCID})

Authors:
\itemize{
  \item Yves Van de Peer \email{yves.vandepeer@psb.vib-ugent.be} (\href{https://orcid.org/0000-0003-4327-3730}{ORCID})
}

}
\keyword{internal}
