% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_and_modules}
\alias{genes_and_modules}
\title{Genes and modules for a coexpression network inferred from SRP201971}
\format{
A 2-column data frame with the following variables:
\itemize{
\item{\emph{Genes} Character, gene ID.}
\item{\emph{Modules} Character, module name.}
}
}
\usage{
data(genes_and_modules)
}
\description{
Expression data (in TPM) and sample metadata were obtained from
Refine Bio. The signed gene coexpression network was inferred with
the \strong{BioNERO} package.
}
\examples{
data(genes_and_modules)
}
\keyword{datasets}
