% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ME}
\alias{ME}
\title{Module eigengenes for a coexpression network inferred from SRP201971}
\format{
A data frame with samples in rows and module eigengenes in columns.
}
\usage{
data(ME)
}
\description{
Expression data (in TPM) and sample metadata were obtained from
Refine Bio. The signed gene coexpression network was inferred with
the \strong{BioNERO} package.
}
\examples{
data(ME)
}
\keyword{datasets}
